Add api wrapper functions to apiWrapper.js and format classes to be in row format
