+++ client/src/css/ClassCard.scss
@@ -1,7 +1,7 @@
 .class-card {
   flex: none;
   width: 175px !important;
-
+  margin: 10px !important;
   .card-content-container {
     display: flex;
     justify-content: space-evenly;
