+++ client/src/css/Profile.scss
@@ -10,5 +10,13 @@
   }
 
 .classes-display{
-  align-content: center;
+  display:flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  column-gap: 10px;
+  flex-wrap:wrap;
+  padding-right: 200px;
+  padding-left: 200px;
 }
+
